requirements.txt
CherryPy
mysql-connector-python
